+++ src/app/_hook/api/votes/useVoteListData.ts
@@ -1,7 +1,7 @@
 import { useMemo } from 'react'
 import { InfiniteData, useSuspenseInfiniteQuery } from '@tanstack/react-query'
 import { PagesResponse } from '@/app/_types/vote.type'
-import { voteKeys } from './votes'
+import { voteKeys } from '.'
 
 interface VoteQueryParams {
   pageParam: string | null
